[fix] 마이페이지와 이용내역 페이지 분리
